api/app/requirements.txt
fastapi
psycopg2-binary
uvicorn
python-dotenv
PyJWT
redis
boto3
requests
pycognito
cryptography
pydantic
email-validator